Avoid "No response from" popup error
Finnally. This was annoying.

When the board is first accessed, if the user navigates away before
any cards are laoded, and then comes back, it would check for updates to
the cards, but the bzBoardLoadTime hadn't been set yet so it was failing.

No response from http://bugzilla.msec.local for request &last_change_time=