Fix for #80 multiple login popups
Some instances of Bugzilla are private and require a login to perform any
kind of query. Bzkanban was designed for public instances so it does
not handle these private instances well. For example on startup it sends
off many queries and these all come back with a 410 error. Handle these by
displaying a user login prompt on the first 410 error we get.

Note, an alert is commented out now. This may supress some useful error and
warning messages that we were getting from Bugzilla for free previously.
For example when dragging a card into the wrong column.